+++ extensions/vendors/illuminate/contracts/Foundation/Application.php
@@ -23,7 +23,6 @@ interface Application extends Container
     /**
      * Get or check the current application environment.
      *
-     * @param  mixed
      * @return string
      */
     public function environment();
@@ -56,7 +55,7 @@ interface Application extends Container
      * Register a deferred provider and service.
      *
      * @param  string  $provider
-     * @param  string  $service
+     * @param  string|null  $service
      * @return void
      */
     public function registerDeferredProvider($provider, $service = null);
@@ -83,13 +82,6 @@ interface Application extends Container
      * @return void
      */
     public function booted($callback);
-
-    /**
-     * Get the path to the cached "compiled.php" file.
-     *
-     * @return string
-     */
-    public function getCachedCompilePath();
 
     /**
      * Get the path to the cached services.php file.